+++ app/controllers/application_controller.rb
@@ -1,5 +1,8 @@
 class ApplicationController < ActionController::Base
   # before_action :set_board, only: [:show]
   # before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
-   
+   def current_user
+    ActiveDecorator::Decorator.instance.decorate(super) if super.present?
+    super
+   end
 end
