+++ app/decorators/board_decorator.rb
@@ -1,0 +1,9 @@
+module BoardDecorator
+  def display_created_at
+    I18n.l(self.created_at, format: :default)
+  end
+
+  def author_name
+    user.display_name
+  end
+end
