+++ app/models/board.rb
@@ -30,18 +30,7 @@ class Board < ApplicationRecord
   # belongs_to :genre, optional: true
 
 
-  def display_created_at
-    I18n.l(self.created_at, format: :default)
-  end
-
-  # def has_written?(board)
-  #   boards.exists?(id: board.id)
-  # end
-
-
-  def author_name
-    user.display_name
-  end
+  
 
   def avatar_image
     if profile&.avatar&.attached?
