+++ app/models/user.rb
@@ -48,13 +48,7 @@ class User < ApplicationRecord
   #   comments.exists?(id: comment.id)
   # end
 
-  def display_name
-    if profile && profile.nickname
-      profile.nickname
-    else  
-      self.email.split('@').first
-    end
-  end
+  
 
   def follow!(user)
     # if user.is_a?(User)
@@ -86,13 +80,7 @@ class User < ApplicationRecord
     profile || build_profile
   end
 
-  def avatar_image
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
-  end
+  
 
   private
   def get_user_id(user)
