--- app/views/boards/_tabs.html.haml
@@ -1,4 +1,0 @@
--# .tabs
--#   .tabs_item.active
-%h2 All Boards 
-
