+++ app/views/boards/index.html.haml
@@ -1,4 +1,5 @@
 .boards
+  = render 'commons/tabs'
   .container
     %h2 All Boards 
     - @boards.each do |board|
