+++ app/views/timelines/show.html.haml
@@ -1,0 +1,7 @@
+Timeline
+.boards
+  = render 'commons/tabs'
+  .container
+    %h2 All Boards 
+    - @boards.each do |board|
+      = render 'commons/board', board: board
